move classes to correct packages
