+++ webapi/src/main/scala/org/knora/webapi/slice/admin/domain/service/KnoraProjectRepo.scala
@@ -8,7 +8,7 @@ import zio.Task
 
 import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectIdentifierADM
 import org.knora.webapi.slice.admin.domain.model.KnoraProject
-import org.knora.webapi.slice.common.service.Repository
+import org.knora.webapi.slice.common.repo.service.Repository
 import org.knora.webapi.slice.resourceinfo.domain.InternalIri
 
 trait KnoraProjectRepo extends Repository[KnoraProject, InternalIri] {
