+++ webapi/src/main/scala/org/knora/webapi/slice/common/api/RestPermissionService.scala
@@ -1,14 +1,14 @@
-package org.knora.webapi.slice.admin.api.service
+package org.knora.webapi.slice.common.api
 
 import zio._
 import zio.macros.accessible
 
 import dsp.errors.ForbiddenException
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
-import org.knora.webapi.slice.admin.api.service.RestPermissionService.isActive
-import org.knora.webapi.slice.admin.api.service.RestPermissionService.isSystemAdmin
-import org.knora.webapi.slice.admin.api.service.RestPermissionService.isSystemOrProjectAdmin
 import org.knora.webapi.slice.admin.domain.model.KnoraProject
+import org.knora.webapi.slice.common.api.RestPermissionService.isActive
+import org.knora.webapi.slice.common.api.RestPermissionService.isSystemAdmin
+import org.knora.webapi.slice.common.api.RestPermissionService.isSystemOrProjectAdmin
 
 /**
  * Provides methods for checking permissions.
